+++ app/plugins/socket.client.ts
@@ -1,0 +1,5 @@
+import { socket } from '~/composables/useSocket'
+
+export default defineNuxtPlugin(() => {
+  socket.connect()
+})